+++ src/Propulsion.Cosmos/Infrastructure.fs
@@ -1,4 +1,8 @@
-﻿namespace Propulsion.Cosmos
+﻿#if COSMOSSTORE
+namespace Propulsion.CosmosStore
+#else
+namespace Propulsion.Cosmos
+#endif
 
 open System
 open System.Threading
@@ -11,7 +15,7 @@ open System.Threading.Tasks
 module private AsyncHelpers =
     type Async with
         /// Asynchronously awaits the next keyboard interrupt event
-        static member AwaitKeyboardInterrupt () : Async<unit> = 
+        static member AwaitKeyboardInterrupt () : Async<unit> =
             Async.FromContinuations(fun (sc,_,_) ->
                 let isDisposed = ref 0
                 let rec callback _ = Task.Run(fun () -> if Interlocked.Increment isDisposed = 1 then d.Dispose() ; sc ()) |> ignore